De-duplicate aggregator code
