+++ packages/validator/src/util/index.ts
@@ -1,4 +1,3 @@
-export * from "./aggregator";
 export * from "./clock";
 export * from "./error";
 export * from "./url";
